+++ src/wasds150/plans/__init__.py
@@ -11,15 +11,19 @@ from typing import Dict, List
 
 from wasds150.models.plan import ChannelPlan
 from wasds150.plans.ftx1_local import FTX1_LOCAL
+from wasds150.plans.ftx1_scan import FTX1_SCAN
 from wasds150.plans.ftx1_wa import FTX1_WA
 from wasds150.plans.h9_ozette import H9_OZETTE
 from wasds150.plans.thd75_ames_lake import THD75_AMES_LAKE
+from wasds150.plans.thd75_scan import THD75_SCAN
 
 _REGISTRY: Dict[str, ChannelPlan] = {
     H9_OZETTE.id: H9_OZETTE,
     FTX1_WA.id: FTX1_WA,
     FTX1_LOCAL.id: FTX1_LOCAL,
+    FTX1_SCAN.id: FTX1_SCAN,
     THD75_AMES_LAKE.id: THD75_AMES_LAKE,
+    THD75_SCAN.id: THD75_SCAN,
 }
 
 
@@ -43,9 +47,11 @@ def get_plan(plan_id: str) -> ChannelPlan:
 
 __all__ = [
     "FTX1_LOCAL",
+    "FTX1_SCAN",
     "FTX1_WA",
     "H9_OZETTE",
     "THD75_AMES_LAKE",
+    "THD75_SCAN",
     "get_plan",
     "list_plans",
     "plan_ids",